Move client opts out of repo
This should aid readability
