+++ opts_test.go
